Update to libarchive v.3.6.0

+++ qlZipInfo/libarchive/archive_string.c
@@ -745,7 +745,7 @@ archive_string_append_from_wcs_in_codepage(struct archive_string *as,
 				dp = &defchar_used;
 			count = WideCharToMultiByte(to_cp, 0, ws, wslen,
 			    as->s + as->length,
-			    (int)as->buffer_length - as->length - 1, NULL, dp);
+			    (int)as->buffer_length - (int)as->length - 1, NULL, dp);
 			if (count == 0 &&
 			    GetLastError() == ERROR_INSUFFICIENT_BUFFER) {
 				/* Expand the MBS buffer and retry. */
@@ -3014,7 +3014,7 @@ archive_string_normalize_C(struct archive_string *as, const void *_p,
 		while ((n2 = parse(&uc2, s, len)) > 0) {
 			uint32_t ucx[FDC_MAX];
 			int ccx[FDC_MAX];
-			int cl, cx, i, nx=0, ucx_size;
+			int cl, cx, i, nx = 0, ucx_size;
 			int LIndex,SIndex;
 			uint32_t nfc;
